aquila 26/aquila_trace/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aquila-trace"
version = "1.0.0"
description = "Multi-layered Intelligence Platform for Financial Crime and Terrorism Financing Detection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AquilaTrace Team", email = "team@aquilatrace.com"}
]
keywords = ["financial-crime", "terrorism-financing", "machine-learning", "graph-neural-networks", "nlp", "blockchain-analytics"]

dependencies = [
    # Core ML libraries
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "pytorch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    
    # Graph Neural Networks
    "torch-geometric>=2.3.0",
    "dgl>=1.1.0",
    "networkx>=3.1",
    
    # NLP
    "transformers>=4.30.0",
    "torch-hub>=0.1.0",
    "spacy>=3.6.0",
    "fasttext>=0.9.2",
    "sentence-transformers>=2.2.0",
    "huggingface-hub>=0.16.0",
    "nltk>=3.8.0",
    
    # Anomaly Detection
    "isolation-forest>=1.0.0",
    "hdbscan>=0.8.0",
    
    # Data Processing & Feature Engineering
    "featuretools>=1.24.0",
    "category-encoders>=2.6.0",
    
    # Blockchain Analysis
    "web3>=6.0.0",
    "requests>=2.31.0",
    
    # API & Server
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "pymongo>=4.5.0",
    
    # Logging & Monitoring
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.18.0",
    
    # Utilities
    "click>=8.1.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "joblib>=1.3.0",
    "hashlib>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
]

docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/aquilatrace/aquila-trace"
Documentation = "https://aquila-trace.readthedocs.io"
Repository = "https://github.com/aquilatrace/aquila-trace.git"
Issues = "https://github.com/aquilatrace/aquila-trace/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
